requirements.txt
pandas>=2.2
numpy>=1.24
httpx>=0.27
python-dotenv>=1.0
apscheduler>=3.10
pyarrow>=14.0
matplotlib>=3.8
flask>=3.0
trading-ig>=0.0.22
lightstreamer-client-lib>=1.0
munch>=4.0
